+++ src/libaktualizr/primary/initializer.cc
@@ -121,7 +121,8 @@ InitRetCode Initializer::initTlsCreds() {
   std::string pkey;
   std::string cert;
   std::string ca;
-  StructGuard<BIO> device_p12(BIO_new_mem_buf(response.body.c_str(), response.body.size()), BIO_vfree);
+  StructGuard<BIO> device_p12(BIO_new_mem_buf(response.body.c_str(), static_cast<int>(response.body.size())),
+                              BIO_vfree);
   if (!Crypto::parseP12(device_p12.get(), "", &pkey, &cert, &ca)) {
     LOG_ERROR << "Received a malformed P12 package from the server";
     return InitRetCode::kBadP12;
